added pie charts and refactored some things
