+++ main.cpp
@@ -1,5 +1,5 @@
-#include "mainwindow.h"
-#include "dbservice.h"
+#include "windows/mainwindow.h"
+#include "services/dbservice.h"
 #include <QApplication>
 #include <QTextCodec>
 #include <QtCharts/QtCharts>
@@ -7,7 +7,6 @@ QT_CHARTS_USE_NAMESPACE
 int main(int argc, char *argv[])
 {
     QApplication a(argc, argv);
-//a.setQuitOnLastWindowClosed(false);
     MainWindow w;
     w.show();
