+++ util/utilenums.h
@@ -1,0 +1,14 @@
+#ifndef UTILENUMS_H
+#define UTILENUMS_H
+
+
+class UtilEnums
+{
+public:
+    UtilEnums();
+    enum Interval {DAY, MONTH, YEAR};
+    enum Type {NONE, INCOMES, EXPENSES, BOTH};
+    enum ChartType {INCOME, EXPENSE, BALANCE};
+};
+
+#endif // UTILENUMS_H
