+++ windows/categoryform.cpp
